in test _multi_add_child_nonroot_sorted, nodes must be reloaded
Because of the way treebeard does raw queries, some operations need to reload a
node after it was modified. This isn't a problem in most views, but some
complex ones that do a lot of operations can be problematic because a raw query
won't update the node object in memory.

This is mainly a documentation bug. The documentation should mention what
queries are safe.