Support models that use multi-table inheritance.
If a model implementing tree behaviour is subclassed, it is possible for the tree to contain a mixture of nodes which are instances of the subclass, and nodes that aren't. Consequently, when we call methods such as get_children, we need to ensure that the queries are performed on the base model rather than the node's own class.