Search request context processor in settings.TEMPLATES.
It is necessary to use different code path for Django 1.0 and newer versus older
versions because 1.10 made the settings.TEMPLATES list mandatory. For older
version, the simple search inside settings.TEMPLATE_CONTEXT_PROCESSORS is kept
in place but for 1.10 and beyond a more complex search inside the list of
template backends and their options is necessary.

This works by checking three conditions for each entry in settings.TEMPLATES:
 * If the backend is `django.template.backends.django.DjangoTemplates`
 * If the `APP_DIRS` flag is `True`
 * If the `django.template.context_processors.request` processor is active
When all three conditions are met, it is assumed that `request` is already
present in the templates at runtime.