Simplify pytest configuration
I managed to replace bits and bobs of code by something hopefully
simpler and more readable. That was mostly done as an exercise to get my
head round what was there, and to allow for dead-simple running of tests
with just specifiying:

$ pytest

or optionally setting an environment variable to run against MySQL or
Postgres as a database backend. Before, you used to be able to set
username and password for those, but until nobody asks for it, I'm not
convinced devs will actually run this locally. We can always reintroduce
that complexity then.