Update mp_tree.py
I added `include_self=False` to `get_descendants()` as the 2nd argument so that we can choose to include the node itself or not. *I switched from Django-mptt to Django-treebeard because Django-mptt is currently unmaintained but Django-treebeard doesn't have `include_self=False` in `get_descendants()` which Django-mptt has. `include_self=False` is really useful.